#12256 Improve JVM Arch Detection
Notify the user if the JVM architecture could not be detected, instead of silently defaulting to 32-bit. Also add a max heap parameter. On machines with RAM sizes in the order of 64GB plus, the initial heap could be multi-gigabyte. This can lead to an out of memory error when checking for the version.


Former-commit-id: 67821ee62aa5b46aa96e6c67d218a7ec69f7892a