Java api: don't mutate fields in SimpleQueryStringBuilder during toQuery execution
We used to mutate the map of fields and weights while transforming to lucene query (toQuery method), by adding the default field if the map is empty. That is an anti-pattern, we should leave the original query as is instead, but still generate the proper lucene query. Also took the chance to fix a couple of issues in SimpleQueryStringBuilderTest#doAssertLuceneQuery.


Former-commit-id: 11b41f1d6861a62458dee412834b8505e30744e4