[test] Remove esoteric apt-get in Vagrantfile
Removes an esoteric `apt-get update` variant used in Vagrantfile that was
causing only parts of the apt repository to update. That was the point of
the command but when it would leave the repository only half built which
made installing anything but Java difficult. The speed isn't worth the
complexity.


Former-commit-id: c427b3e9c2201d520d941e682c40ce7148fdaf55