Be more careful with RoutingTable.Builder
RoutingTable.Builder#build can only be called once but didn't have any
checks to make sure it _was_ only called once. And the error message it
threw when called more than once was a NullPointerException. This makes
it throw IllegalStateException when you try to reuse the builder in any way.


Former-commit-id: 56318df10ce35bb778ed1b99012e3a0093685f0d